+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [3.9.1]
-- Long press share button to share cURL command
+- Added `quickShareAction` property. This action will be called on long press on share button. It has AliceHttpCall as argument and should be used to handle share action. By default it will share cURL command.
 
 ## [3.9.0]
 - Bump dependencies